Add mangling for invalid identifier strings.
If a C reserved word, for example, short or long, is used as an
identifier in carp, the resultant C code contains an invalid identifier
and fails to compile. This change adds a function for mangling reserved
words into valid C identifiers and includes two initial terms, short and
long.

We apply the string mangling after symbol/character mangling as the
string resulting from a successful character mangle is already a valid
identifier even if it contains a reserved word.