Macros: Add unit macro.
Adds a unit macro, a convenience macro for returning () in
side-effecting functions. This is useful, for example, for match forms
in which only one match should result in a side effect. The other
matches still require forms for completeness, but they should simply
return (). Unit might be easier to read in these cases, as parens can
get lost in the sea of surrounding parens in the definition.