Concretize: Fix top level lambda naming
This change adjusts the way in which function nest levels are
incremented to ensure top level lambdas are assigned the correct nest
level. Top level lambdas used to use their expressions as names,
resulting in invalid c. A check on nest levels in the lambda naming code
now prevents this.