Add support for explicit null pointers
This change adds a simple function to create a *null* void pointer from
Carp.

The impetus for this is that older C libraries sometimes use null
pointers as arguments to functions; I encountered this in practice while
working with ncurses; many of the attr functions in ncurses take a void
pointer argument which was "reserved for future use" and never actually
used. The current implementation of ncurses I have expects an Int
pointer for two of these functions but otherwise checks that this
argument is NULL (according to the curs_attr) man page.

I couldn't find a simple way to send a null pointer from Carp, so I
added this function.