Fix non-exhuastive patterns in concretize
This change supports more flexible Sumtype declarations.

We used to allow a sumtype declaration such as:

`(deftype (State a) Done (Value [a]))`

Which the compiler accepts. However, concretization for this type
failed, since we didn't handle the solitary Sym case `Done` when
concretizing the type.

This commit fixes that issue, so that now the following works:

```
(deftype (State a) Done (Value [a]))
;; Omitting the type annotation here would result in the obvious
;; unresolved generic type error.
(defn emit-done [] (the (State Int) (State.Done)))
(emit-done)
;; => State.Done
```