Make lookups generic, recursively lookup impls
Before, we were only looking up interface implementation in a single
context. We needed to use something like MultiSymLookup, but on the
presence and value of a meta key, rather than a binder name.

So, I extracted the local fns in multiSymLookup out into their own
functions and make lookup functions generic, and of the type:

LookupFunc :: a -> Env -> [Binders]

This seems to suit most purposes. Eventually, MultiSymLookup can be
redefined in terms of these simpler lookup mechanisms.