Update `sigFromDefnOrDef` to handle relative definitions
Previously, sig resolution for definitions inside of modules didn't work
because this function didn't check for relative paths. This meant that
pretty much all signatures weren't honored, since forward declarations
of metadata only work within the context of modules, and so are
unqualified paths (the right module qualifications need to be extracted
from the Context).