Fix signature of Control.iterate-until
Now that sigs work, turns out the signature for iterate-until and it's
definition don't match.

There was an extra argument in the sig, which I've removed. Adding a ref
to `result` also confused the type checker, since result technically has
two types in the definition whenever `b` is not a ref. I believe the
side effect of this change is that the `start` argument must now be a
`Ref`, but at least it's well-typed.