Add s-exp primitive
This primitive returns the s-expression associated with a binding--in
other words, the form a user wrote to define a binding.

We handle special cases such as interfaces and deftypes, whose original
defining forms are stored in the type environment instead of the global
environment.

Note that presently this doesn't always return *exactly* the form the
user wrote. Specifically, the following cases are special:

- Interfaces -- returns `(interface foo)` instead of the full form
- Defndynamic -- returns `(dynamic foo [x] x)` instead of `defndynamic`
- Defmacro -- returns `(macro [x] x)` instead of `defmacro`

In spite of these wrinkles, I think it's a useful function to have,
particularly for type introspection in dynamic functions/macros.