Map special forms to symbols
This commit makes the forms returned by the s-exp primitive usable with
other dynamic functions that operate on symbols. By default, a special
form like `Defn` is *not* a symbol. So, we map all the special forms to
symbols in order to use the returned s-exps smoothly with other dynamic
functions (which operate on symbols).