Make Sexpression a command, enhance module introspection
Originally, s-expr was a primitive and didn't evaluate its arguments.
This however make it difficult to use, in addition to forcing us to make
an extra unnecessary lookup. This commit converts s-expr to a command,
which makes it a lot simpler to use in macro definitions and in dynamic
code in general.

Additionally, I've enhanced it to do two things:

- Allow users to get the module for a type instead of the definition
  form using an optional argument.
- List the contents of a module upon getting its s-expr.

Many thanks to @hellerve who both suggested this change and advised.