Restore tyvar check in constraints; special case interface
Previously, I'd removed the type var check in constraint conflict
checking. This doesn't seem to be problematic, but it might not be the
best way to implement the desired result, which is special conflict
checking for interfaces.

This approach is more conservative and simply bypasses the check for
interface implementation constraint solving only.