Type check `set!` calls
This change enforces that `set!` calls do not change a variable's type.
Attempting to set a variable of type T to a value of type W will result
in an error. Dynamic variables are not type checked and may be set to
any value.

Generally speaking, it's safer to enforce this restriction on variables
for soundness reasons, and it's often the behavior programmers desire.
Additionally, this check resolves a problem whereby, previously,
reassigning a variable to a value of a different type and using it in
code thereafter would cause the compiler to hang.

Here's a brief illustration of the new behavior:

```
(def x 1)
(set! x "foo")
=> can't `set!` x to a value of type (Ref String a), x has type Int at
REPL:1:7.
(set! x 2)
(Int.+ 1 x)
=> 3
;; Dynamics can still be set to anything
(defdynamic y 2)
(set! y "foo")
:i y
(defdynamic y "foo") : Dynamic
;; Of course, one can still *redefine* a variable to change its type.
(set! x 3.4)
=> can't `set!` x to a value of type Double, x has type Int at REPL:1:7.
(def x 0.0)
[WARNING] Definition at line 1, column 1 in 'REPL' changed type of 'x'
from Int to Double
(set! x 10.0)
:i x
=> x : Double
```