Refactor set! slightly; allow setting the global env locally
This change handles a case the previous implementation missed, that is,
when a user attempts to set a global variable in a local environment,
such as a let binding:

```
(def x 1)
(let [y 2] (set! x y))
(Int.+ 1 x)
=> 3
```

Internal definitions have precedence over globals, that is, if another
`x` in the let binding above shadowed the `x` in the global environment,
only the local `x` would be updated:

```
(def x 1)
(let [x 2] (set! x 3))
(Int.+ 1 x)
=> 2
```