Add a global repl? var that indicates if we're running a repl
This commit adds a variable called `repl?` to the starting global
environment that designates whether or not Carp is running as a REPL and
gives Carp programs access to this information. One can use this to
conditionally trigger some logic, e.g. such as printing some
introductory information, when Carp is running as a REPL only.

Here's an example of this behavior:

```
;; Start Carp in REPL mode.
carp
(if repl? 1 2)
=> 1
```

```
;; source for foo.carp
(defn main []
  (if repl? 1 2))
;; Run foo
carp -x foo.carp
=> exited with return value 2
```