Expose Carp's execution mode in the Introspect module
This is an extension of the previous commit. Rather than only give users
insight into whether Carp is running as a REPL, we expose variables at
the static and dynamic level that report the current execution mode Carp
is running in.

Using these variables, users can conditionally trigger logic based on
the compilers execution more.