Only expose execution modes when core is enabled
The ExecutionMode type relies on String, which is defined in core, and
the `doc` invocations also rely on core. Adding the execution mode
types/vars without core leads to breakages (since these dependencies are
missing) so we now check that core is enabled before adding these to the
starting environment.