Eval: Add Address to eval
The `address` keyword is useful when one needs to circumvent Carp's
memory management for particular cases. However, the evaluator
previously didn't understand `address` in most contexts, so it proved
difficult to toy around with in the REPL.

This commit adds basic support for calling `address` in the REPL by
adding a case to the evaluator. This allows users to test out functions
that take `Ptr` arguments such as `Pointer.to-value`. The `address`
keyword only accepts a symbol as an argument (since getting the address
of a literal value doesn't always make sense (e.g. &2) and produces
invalid C):

    (def a 1)
    (address a)
    => a
    (type (address a))
    => (Ptr Int)
    (Pointer.to-value (address a))
    => 1

Note: one will still typically need to wrap it in a function in
higher-order contexts.